Sanitize unnecessary usage of double

+++ WinQuake/cl_demo.c
@@ -325,13 +325,13 @@ CL_FinishTimeDemo
 void CL_FinishTimeDemo (void)
 {
 	int		frames;
-	float	time;
+	double	time;
 	
 	cls.timedemo = false;
 	
 // the first frame didn't count
 	frames = (host_framecount - cls.td_startframe) - 1;
-	time = realtime - cls.td_starttime;
+	time = realtime - (double)cls.td_starttime;
 	if (!time)
 		time = 1;
 	Con_Printf ("%i frames %5.1f seconds %5.1f fps\n", frames, time, frames/time);
